fix(types): remove plugin interfaces from index.d.ts and co-locate with implementations
- Removed LanguagePlugin, ThemePlugin, StoragePlugin, and AppStatePlugin from index.d.ts
- Moved each interface to its respective plugin .d.ts file
- Updated src/plugins/index.d.ts to import from individual plugin files
- Removes circular dependency pattern
- Completes explicit imports migration from PR 555
