+++ src/plugins/storage/storage.d.ts
@@ -15,10 +15,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { StoragePlugin } from '@lightningjs/blits'
-
-// Re-export StoragePlugin for direct imports
-export type { StoragePlugin }
+export interface StoragePlugin {
+  get<T = unknown>(key: string): T | null
+  set(key: string, value: unknown): boolean
+  remove(key: string): void
+  clear(): void
+}
 
 declare const storage: {
   readonly name: 'storage'
