+++ src/plugins/theme.d.ts
@@ -15,16 +15,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { ThemePlugin } from '@lightningjs/blits'
+export interface ThemePlugin {
+  get<T = unknown>(key: string): T | undefined
+  get<T>(key: string, fallback: T): T
+  set(theme: string): void
+}
 
 export interface ThemePluginConfig {
   themes?: Record<string, Record<string, unknown>>
   current?: string
   base?: string
 }
-
-// Re-export ThemePlugin for direct imports
-export type { ThemePlugin }
 
 declare const theme: {
   readonly name: 'theme'
